refactor(anime): split API utilities
